Conversion to using ASSIMP start + GitIgnore Change
We've started work on our conversion to using ASSIMP to load models (Referencing the materials found on LearnOpenGL) as ASSIMP allows us to get data such as tangent and bitangent information from the models, along with having more complex models.

We've mostly gotten it working, though there is plenty of temporary testing code, and for some reason between a model's materials being loaded and the model finishing construction we lose access to them. For now we're resolving this by reloading the materials after the model is constructed, but we should try to find a better way to solve this.

Furthermore, we've changed the .gitIgnore to also include the elements of the 'VisualStudio.gitignore' found on GitHub, in addition to what we already had for our c++ .gitignore.