Quaternions, Camera Transform, & Rotations
We've changed the rotations over to using quaternions from euler angles, and have changed the Camera to use a transform rather than having its own logic for position and rotation.

Additionally, we've added a function that allows us to rotate around ourselves given a specified axis in either local or world space.