.gitIgnore Required .dll Files Update
We've updated the .gitIgnore so that we no longer ignore our required .dll files (For glew, assimp, & SDL).
We also updated the removal for the .gitIgnore to get this to work, which has removed some files which were included in the .gitIgnore but were missed by the purge when we were first updating it.