Skybox Cleanup & Cubemap Returns Pointer
We've cleaned up the skybox code a little and moved it back into the 'Skybox' class.

We've also figured out that one of the issues which was causing the skybox to not display (In addition to using the wrong function for drawing the cube) is that something was going wrong when getting a unique pointer for the Texture returned by a Cubemap (This also happened with regular pointers). We've changed the function to return a pointer, which now lets us work with it.
We are however currently setting a std::unique_ptr directly with a pointer, as we are unsure whether std::make_unique with a pointer passed would duplicate the texture or cause a memory leak.