Texture Copy Constructor Removal + Unique/Shared Ptr Creation Fixes
We've removed the ability to access the copy constructor of the 'Texture' class, and through doing so have learned that passing an already existing instance of a class into make_unique/make_shared copies that instance. We've fixed the occurrences of this that we could find so those objects are no longer being copied.