Skybox Blend, Day/Night Sky Textures, & Cubemap from Single Image
We've implemented the blending between our new Day & Night skyboxes, based on the current progress through the day in the same way that our sunlight colour is being evaluated.

Additionally, we've implemented the ability to create a cubemap texture from a single image laid out in the horizontal cross layout.