PBR Directional Lights & LightingTests.frag Cleanup
We've implemented the PBR code for Directional Lighting (After the initial calculations it is the same as for Point Lights, so it is now a function).

Additionally, we've cleaned up the 'LightingTests.frag' shader, renaming some variables and shifting some comments around.
We still don't fully know what 'H' represents, but are pretty sure that our guess of 'height' was wrong, so we've reverted that naming change.