Normal Mapping
We've implemented Normal Mapping, fixing the base 'NormalMapping' shader (We were missing a transpose() call when calculating the TBN matrix) and implementing normal mapping in the 'LightingTests' shader.

Note: For some reason the 'LightingTests' shader doesn't draw if called before another shader (It needs to be called last). We should try and figure out why.