Low Poly Tree & No Normal Map Rendering Option
We've added in a Low Poly Tree model which we are going to scatter throughout the scene to make it feel less empty.

Additionally, we've added an option to the LightingTests shader to allow us to draw without using the normal map, for cases such as when we have none.