Shader Lighting Parameter Updating Improvement
We've improved the way that we're setting the lighting parameters in our shaders by having it that the lights set the required values again, reducing code duplication.

Furthermore, we've made it so that no errors are thrown if a shader doesn't have all the parameters, allowing us to pass in shaders that only require a lights position, for example.