Local Position Setting, Light Transforms, & Child Rotation Fix
We've implemented the ability to set the local position of our objects (Which takes into account their parent's rotation and scale too), along with giving lights transforms and having their references be children of them rather than setting their positions manually.

Additionally, we've fixed a mistake in our rotation calculation code which was meaning that our child rotations were being incorrectly calculated when the rotation was along an axis which didn't align with the child.