More Tree Instances, Model Placement, & Minor Code Cleanup
We've added in the generation of more trees in the scene, along with updating their position, rotation, and scale values, along with the values for the bench and lanterns.

We've also done a little code clean-up, and have removed the backpack test model that was still in the resources folder.