Moon Light & Lantern Light Diffuse Change
We've added a second directional light to represent the moon, along with altering the diffuse colour of the point lights being used in the lanterns to make them more yellow-green in colour, similar to actual lantern light.

We also doubled the day length, but have subsequently also doubled the time speed-up value when holding 'Q'.